вынес логику комментариев в отдельный файл
