+++ js/big-photo.js
@@ -1,3 +1,5 @@
+import {renderComments} from './comments.js';
+
 const bigPicture = document.querySelector('.big-picture');
 const likesCount = bigPicture.querySelector('.likes-count');
 const bigPictureImg = bigPicture.querySelector('.big-picture__img img');
@@ -5,13 +7,8 @@ const commentsCountContainer = bigPicture.querySelector('.social__comment-count'
 const commentsCount = commentsCountContainer.querySelector('.comments-count');
 const cancelBtn = document.querySelector('.big-picture__cancel');
 const description = bigPicture.querySelector('.social__caption');
-const socialComments = bigPicture.querySelector('.social__comments');
 const commentsLoader = document.querySelector('.comments-loader');
 const body = document.querySelector('body');
-const commentTemplate = document.querySelector('#comment')
-  .content
-  .querySelector('.social__comment');
-
 
 cancelBtn.addEventListener('click',() => {
   bigPicture.classList.add('hidden');
@@ -25,15 +22,6 @@ body.addEventListener('keydown', (evt) => {
   }
 });
 
-const createCommentElement = ({avatar, message, name}) => {
-  const comment = commentTemplate.cloneNode(true);
-  comment.querySelector('.social__picture').src = avatar;
-  comment.querySelector('.social__picture').alt = name;
-  comment.querySelector('.social__text').textContent = message;
-
-  return comment;
-};
-
 export const showBigPhoto = (photo) => {
   bigPicture.classList.remove('hidden');
   bigPictureImg.setAttribute('src', photo.url);
@@ -41,10 +29,7 @@ export const showBigPhoto = (photo) => {
   description.textContent = photo.description;
   commentsCount.textContent = photo.comments.length;
 
-  socialComments.innerHTML = '';
-
-  const commentElements = photo.comments.map((comment) => createCommentElement(comment));
-  socialComments.append(...commentElements);
+  renderComments(photo.comments);
 
   commentsCountContainer.classList.add('hidden');
   commentsLoader.classList.add('hidden');
